JSON RPC server of Workflow Analysis, progress & error reporting
This change improves a couple of things:
  * Switch to LSP4J to create a proper JSON RPC server for the workflow
analysis. This makes the code more clear (interfaces for communication)
and eases progress and error reporting and opens opportunities for
bi-directional communciation tomorrow.
  * Use JsonRpcProxies to simplify communicating with the server from
within the front-end
  * Allow to start the workflow server from the IDE and use the instance
started from the IDE in the Theia app for debugging. Therefore, you just
need to provide a port as a program argument for the Theia backend. If
this argument exists, we don't start a workflow analysis server on our
own, but will just connect to an existing one. See also
.vscode/launch.json
  * Shows a progress message while generating code and while analyzing a
workflow. The progress message will disappear once the action is
completed or an error occurs.
  * Shows errors and details of errors that may happen during the
workflow analysis, starting the workflow analysis server, etc.