JsonForms Tree - Extend node factory to support property keys
The JsonForms tree's node factory interface only supported number indices to define
where in a property a new tree node's data is located.
While this is sufficient for data stored in an array, it doesn't support objects.

For this reason the index was renamend to indexOrKey and the type extented to number | string.