Adapt tree editor for usage of IDs, addition of nodes and task rename
Changing of the modelserver backend requires adaption of the tree
editor. This commit does not cover all cases. The following ones are
handled:

* Addition of new tasks, nodes, edges in glsp are handled in the tree
* Task rename command
* Setting of direct primitive properties

On the model server side:
Introduce a new codec that encodes the xmi ids into the delivered json
data in property "@id"

Missing:
* Handling of custom compound commands for remove
* Using custom compound commands for remove by tree (currently normal
emf remove commands)
* Handling of property changes of nested objects in a node (e.g. Display
in ControlUnit)
* Handling of RAM array. This could also partly be caused by other parts
(e.g. jsonforms) as it does not work but no errors are logged anywhere.