Fix validation after reload
Previously the validation would not return correct ids after a reload.
To fix this the validator needed to be updated.
A fix for the modelserver was opened, this is a workaround for now.
This can removed again, once we updated to the modelserver version
containing the fix.

Modelserver change for reference: https://github.com/eclipse-emfcloud/emfcloud-modelserver/pull/244